+++ DFTFringe.pro
@@ -262,6 +262,7 @@ SOURCES += SingleApplication/singleapplication.cpp \
     zapm.cpp \
     zernikedlg.cpp \
     zernikeeditdlg.cpp \
+    zernikepolar.cpp \
     zernikeprocess.cpp \
     zernikes.cpp \
     zernikesmoothingdlg.cpp
@@ -389,6 +390,7 @@ HEADERS += bezier/bezier.h \
     wftstats.h \
     zernikedlg.h \
     zernikeeditdlg.h \
+    zernikepolar.h \
     zernikeprocess.h \
     zernikes.h \
     zernikesmoothingdlg.h
