+++ zernikeprocess.h
@@ -140,36 +140,7 @@ public slots:
 };
 
 //TODO separate class into different files
-class zernikePolar : public QObject
-{
-    Q_OBJECT
-public:
-    explicit zernikePolar(){};
-    static zernikePolar *get_Instance();  //TODO check if we really need a singleton here
-    void init(double rho, double theta, size_t nbTerms = 48);
-    double zernike(size_t z);
-private:
-     static zernikePolar *m_instance;
-     double rho2;
-     double rho3;
-     double rho4;
-     double rho5;
-     double rho6;
-     double rho8;
-     double rho10;
-     double costheta;
-     double sintheta;
-     double cos2theta;
-     double sin2theta;
-     double cos3theta;
-     double sin3theta;
-     double cos4theta;
-     double sin4theta;
-     double cos5theta;
-     double sin5theta;
-     size_t m_nbTermsComputed;
-     double zernTerms[49]; // TODO if no singleton, use correct length
-};
+
 
 void debugZernRoutines();
